backend/requirements.txt
astroid==2.4.2
attrs==20.3.0
bcrypt==3.2.0
boto3==1.14.49
botocore==1.17.49
certifi==2020.6.20
cffi==1.14.2
chardet==3.0.4
click==7.1.2
cryptography==2.8
dnspython==2.0.0
docutils==0.15.2
Flask==1.1.2
gunicorn==20.0.4
idna==2.10
importlib-metadata==3.4.0
iniconfig==1.1.1
isort==5.4.2
itsdangerous==1.1.0
Jinja2==2.11.2
jmespath==0.10.0
lazy-object-proxy==1.4.3
MarkupSafe==1.1.1
marshmallow==3.7.1
mccabe==0.6.1
numpy==1.19.1
packaging==20.8
pandas==1.1.1
pluggy==0.13.1
psycopg2-binary==2.7.5
py==1.10.0
pycparser==2.20
PyJWT==1.7.1
pylint==2.6.0
pymongo==3.11.0
pyparsing==2.4.7
pytest==6.2.1
python-dateutil==2.8.1
python-dotenv==0.14.0
python-http-client==3.3.1
pytz==2020.1
redis==3.5.3
requests==2.24.0
s3transfer==0.3.3
sendgrid==6.5.0
shortuuid==1.0.1
six==1.15.0
starkbank-ecdsa==1.0.0
Tenjin==1.1.1
toml==0.10.1
typed-ast==1.4.1
typing-extensions==3.7.4.3
urllib3==1.25.10
Werkzeug==1.0.1
wrapt==1.12.1
zipp==3.4.0
zxcvbn==4.4.28
